Increase default font size and color
I've found the paragraph text of the current docs very hard to read—I'm nearsighted and the 14px font-size really forces me to get close to see what I'm reading. Upping the size to 16px (or 1rem) is pretty standard for readability purposes, and improves site accessibility for the visually impaired. I also changed to color to a warm gray to give the paragraph text contrast from the headlines so everything is easily distinguishable at first glance.